refactor(ui-editor): extract PlaceholderWidget into standalone file
