+++ packages/ui/ui-editor/src/extensions/tags/index.ts
@@ -8,3 +8,4 @@ export * from './typewriter';
 export * from './xml-block-decoration';
 export * from './xml-formatting';
 export * from './xml-tags';
+export { PlaceholderWidget as XmlPlaceholderWidget, type XmlWidgetNotifier } from './placeholder-widget';
